--- src/render.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import {addPost, updateNewPostText} from './redux/state'
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-
-export let rerenderEntireTree = (state) => {
-    root.render(
-        <React.StrictMode>
-            <App state={state} addPost={addPost} updateNewPostText={updateNewPostText}/>
-        </React.StrictMode>
-    );
-}
-
